Moved old generic parsers
